+++ lib/chess/piece.ex
@@ -1,15 +1,28 @@
-# To re-enable all debug messages in this file, run the following command:
-# sed 's/#IO/IO/g' -i piece.ex
-
 defmodule Chess.Piece do
-  defstruct color: nil,
-            type: nil
+  use Ecto.Schema
+#  import Ecto.Changeset
+#  alias Chess.{Board, Repo, Accounts}
+
+  schema "startconfigurations" do
+    field :type, Ecto.Enum, values: [:rook, :knight, :bishop, :queen, :king, :pawn]
+    field :color, :string
+
+    timestamps(type: :utc_datetime)
+  end
 
   def glyphs() do
     %{:white =>
-        %{:rook => "♖", :knight => "♘", :bishop => "♗", :queen => "♕", :king => "♔", :pawn => "♙"},
+        %{:rook => "♖", :knight => "♘", :bishop => "♗", :queen => "♕", :king => "♔", :pawn => "♙",
+	  :dragon => "🐉", 
+	  :wizard => "🧙‍♂️",
+	  :ninja => "🥷",  
+	  :phoenix => "🦅"},
       :black =>
-        %{:rook => "♜", :knight => "♞", :bishop => "♝", :queen => "♛", :king => "♚", :pawn => "♟"},
+        %{:rook => "♜", :knight => "♞", :bishop => "♝", :queen => "♛", :king => "♚", :pawn => "♟",
+          :dragon => "🐉",
+          :wizard => "🧙‍♂️",
+          :ninja => "🥷",
+          :phoenix => "🦅"}
     }
   end
 end
