+++ lib/chess/piece/moves.ex
@@ -1,194 +1,185 @@
+# To re-enable all debug messages in this file, run the following command:
+# sed 's/#IO/IO/g' -i piece.ex
+
 defmodule Chess.Piece.Moves do
+  require Logger
+
+  # Dragon movements (Queen + Knight)
   def get(%Chess.Board{cells: cells},
-                     %Chess.Piece{color: color, type: :king},
-                     {row, col}) do
-    #IO.puts("\n=== King Move Calculation ===")
-    #IO.inspect({row, col, color}, label: "Calculating get for king")
-
-    # king can move to any one (1) adjacent space
-    get = [
-      {row + 0, col + 1},  # Forward
-      {row + 1, col + 1},  # Forward-right
-      {row + 1, col + 0},  # Right
-      {row + 1, col - 1},  # Back-right
-      {row + 0, col - 1},  # Back
-      {row - 1, col - 1},  # Back-left
-      {row - 1, col + 0},  # Left
-      {row - 1, col + 1},  # Forward-left
-    ]
-
-    valid_get = Enum.filter(get, fn pos ->
-      Map.has_key?(cells, pos) and
-        (cells[pos] == nil or cells[pos].color != color)
-    end)
-
-    #IO.inspect(valid_get, label: "Valid king get")
-    valid_get
+          %Chess.Piece{type: :dragon, color: color},
+          {row, col}) do
+    Logger.info("Calculating Dragon moves from {#{row}, #{col}}")
+    queen_moves = get_queen_moves(cells, {row, col}, color)
+    knight_moves = get_knight_moves(cells, {row, col}, color)
+    moves = Enum.uniq(queen_moves ++ knight_moves)
+    Logger.info("Dragon can move to: #{inspect(moves)}")
+    moves
   end
 
+  # Wizard movements (Teleport anywhere empty or capturable)
   def get(%Chess.Board{cells: cells},
-                     %Chess.Piece{color: color, type: :rook},
-                     {row, col}) do
-    #IO.puts("\n=== Rook Move Calculation ===")
-    #IO.inspect({row, col, color}, label: "Calculating get for rook")
-
-    # Calculate get in each direction
-    directions = [{0, 1}, {0, -1}, {1, 0}, {-1, 0}]  # Up, Down, Right, Left
-    
-    get = Enum.flat_map(directions, fn {row_dir, col_dir} ->
-      find_get_in_direction(cells, {row, col}, {row_dir, col_dir}, color)
-    end)
-
-    #IO.inspect(get, label: "Valid rook get")
-    get
+          %Chess.Piece{type: :wizard, color: color},
+          {row, col}) do
+    Logger.info("Calculating Wizard moves from {#{row}, #{col}}")
+    moves = for r <- 0..7,
+      c <- 0..7,
+    {r, c} != {row, col},
+      cells[{r, c}] == nil || cells[{r, c}].color != color,
+      do: {r, c}
+    Logger.info("Wizard can move to: #{inspect(moves)}")
+    moves
   end
 
+  # Ninja movements (Knight + Adjacent)
   def get(%Chess.Board{cells: cells},
-                     %Chess.Piece{color: color, type: :bishop},
-                     {row, col}) do
-    #IO.puts("\n=== Bishop Move Calculation ===")
-    #IO.inspect({row, col, color}, label: "Calculating get for bishop")
-
-    # Calculate get in diagonal directions
-    directions = [{1, 1}, {1, -1}, {-1, 1}, {-1, -1}]
-    
-    get = Enum.flat_map(directions, fn {row_dir, col_dir} ->
-      find_get_in_direction(cells, {row, col}, {row_dir, col_dir}, color)
-    end)
-
-    #IO.inspect(get, label: "Valid bishop get")
-    get
+          %Chess.Piece{type: :ninja, color: color},
+          {row, col}) do
+    Logger.info("Calculating Ninja moves from {#{row}, #{col}}")
+    knight_moves = get_knight_moves(cells, {row, col}, color)
+    adjacent_moves = get_adjacent_moves(cells, {row, col}, color)
+    moves = Enum.uniq(knight_moves ++ adjacent_moves)
+    Logger.info("Ninja can move to: #{inspect(moves)}")
+    moves
   end
 
+  # Phoenix movements (Jumping diagonals)
   def get(%Chess.Board{cells: cells},
-                     %Chess.Piece{color: color, type: :queen},
-                     {row, col}) do
-    #IO.puts("\n=== Queen Move Calculation ===")
-    #IO.inspect({row, col, color}, label: "Calculating get for queen")
+          %Chess.Piece{type: :phoenix, color: color},
+          {row, col}) do
+    Logger.info("Calculating Phoenix moves from {#{row}, #{col}}")
+    moves = get_diagonal_jumping_moves(cells, {row, col}, color)
+    Logger.info("Phoenix can move to: #{inspect(moves)}")
+    moves
+  end
 
-    # Combine rook and bishop get
+  # King movements (One square in any direction)
+  def get(%Chess.Board{cells: cells},
+          %Chess.Piece{type: :king, color: color},
+          {row, col}) do
+    Logger.info("Calculating King moves from {#{row}, #{col}}")
+    moves = get_adjacent_moves(cells, {row, col}, color)
+    Logger.info("King can move to: #{inspect(moves)}")
+    moves
+  end
+
+  # Pawn movements
+  def get(%Chess.Board{cells: cells},
+          %Chess.Piece{type: :pawn, color: color},
+          {row, col}) do
+    Logger.info("Calculating Pawn moves from {#{row}, #{col}}")
+    direction = if color == :white, do: -1, else: 1
+    start_col = if color == :white, do: 6, else: 1
+
+    # Forward moves
+    forward_moves = 
+    if valid_position?({row, col + direction}) && cells[{row, col + direction}] == nil do
+      moves = [{row, col + direction}]
+      if col == start_col && 
+        valid_position?({row, col + (direction * 2)}) && 
+        cells[{row, col + (direction * 2)}] == nil do
+        [{row, col + (direction * 2)} | moves]
+      else
+        moves
+      end
+    else
+      []
+    end
+
+    # Capture moves
+    capture_moves =
+      [{row - 1, col + direction}, {row + 1, col + direction}]
+      |> Enum.filter(fn pos ->
+      valid_position?(pos) &&
+        cells[pos] != nil &&
+        cells[pos].color != color
+    end)
+
+      moves = forward_moves ++ capture_moves
+      Logger.info("Pawn can move to: #{inspect(moves)}")
+      moves
+  end
+
+  # Helper functions
+  defp get_queen_moves(cells, {row, col}, color) do
     directions = [
-      {0, 1}, {0, -1}, {1, 0}, {-1, 0},  # Rook get
-      {1, 1}, {1, -1}, {-1, 1}, {-1, -1}  # Bishop get
+      {-1, -1}, {-1, 0}, {-1, 1},
+      {0, -1},           {0, 1},
+      {1, -1},  {1, 0},  {1, 1}
     ]
     
-    get = Enum.flat_map(directions, fn {row_dir, col_dir} ->
-      find_get_in_direction(cells, {row, col}, {row_dir, col_dir}, color)
+    Enum.flat_map(directions, fn {dr, dc} ->
+      get_line_moves(cells, {row, col}, {dr, dc}, color)
     end)
-
-    #IO.inspect(get, label: "Valid queen get")
-    get
   end
 
-  def get(%Chess.Board{cells: cells},
-                     %Chess.Piece{color: color, type: :knight},
-                     {row, col}) do
-    #IO.puts("\n=== Knight Move Calculation ===")
-    #IO.inspect({row, col, color}, label: "Calculating get for knight")
-
-    get = [
+  defp get_knight_moves(cells, {row, col}, color) do
+    [
       {row + 2, col + 1}, {row + 2, col - 1},
       {row - 2, col + 1}, {row - 2, col - 1},
       {row + 1, col + 2}, {row + 1, col - 2},
       {row - 1, col + 2}, {row - 1, col - 2}
     ]
-
-    valid_get = Enum.filter(get, fn pos ->
-      Map.has_key?(cells, pos) and
-        (cells[pos] == nil or cells[pos].color != color)
+    |> Enum.filter(fn pos ->
+      valid_position?(pos) && (cells[pos] == nil || cells[pos].color != color)
     end)
-
-    #IO.inspect(valid_get, label: "Valid knight get")
-    valid_get
   end
 
-  # Pawn get - keeping your existing implementation as it works correctly
-  def get(%Chess.Board{cells: cells},
-                     %Chess.Piece{color: color, type: :pawn},
-                     {row, col}) do
-    direction = if color == :white, do: -1, else: 1
-    start_col = if color == :white, do: 6, else: 1
-
-    #IO.puts("\n=== Pawn Move Calculation ===")
-    #IO.inspect({row, col, color}, label: "Calculating get for pawn")
-    #IO.inspect(direction, label: "Movement direction")
-    #IO.inspect(start_col, label: "Starting column")
-
-    # Forward get
-    forward_get = 
-      if Map.has_key?(cells, {row, col + direction}) do
-        #IO.puts("Checking forward to {#{row}, #{col + direction}}")
-        if cells[{row, col + direction}] == nil do
-          one_forward = {row, col + direction}
-          get = [one_forward]
-          #IO.puts("One move forward available: #{inspect(one_forward)}")
-
-          if col == start_col do
-            two_forward = {row, col + (direction * 2)}
-            #IO.puts("Checking two forward to #{inspect(two_forward)}")
-            if Map.has_key?(cells, two_forward) && cells[two_forward] == nil do
-              #IO.puts("Two get forward available")
-              [two_forward | get]
-            else
-              get
-            end
-          else
-            get
-          end
-        else
-          #IO.puts("Forward square is blocked")
-          []
-        end
-      else
-        #IO.puts("Forward square is off board")
-        []
-      end
-
-    capture_positions = [
-      {row - 1, col + direction},  # Capture left diagonal
-      {row + 1, col + direction}   # Capture right diagonal
+  defp get_adjacent_moves(cells, {row, col}, color) do
+    [
+      {row - 1, col - 1}, {row - 1, col}, {row - 1, col + 1},
+      {row, col - 1},                     {row, col + 1},
+      {row + 1, col - 1}, {row + 1, col}, {row + 1, col + 1}
     ]
-
-    #IO.inspect(capture_positions, label: "Checking capture positions")
-
-    capture_get = 
-      capture_positions
-      |> Enum.filter(fn pos ->
-        #IO.inspect(pos, label: "Checking position")
-        if Map.has_key?(cells, pos) && cells[pos] != nil do
-          enemy = cells[pos].color != color
-          #IO.puts("Found #{cells[pos].color} piece, can capture: #{enemy}")
-          enemy
-        else
-          #IO.puts("No capturable piece")
-          false
-        end
-      end)
-
-    #IO.inspect(forward_get, label: "Forward get")
-    #IO.inspect(capture_get, label: "Capture get")
-
-    all_get = forward_get ++ capture_get
-    #IO.inspect(all_get, label: "All valid get")
-    all_get
+    |> Enum.filter(fn pos ->
+      valid_position?(pos) && (cells[pos] == nil || cells[pos].color != color)
+    end)
   end
 
-  # Helper function to find get in a specific direction until blocked
-  defp find_get_in_direction(cells, {row, col}, {row_dir, col_dir}, color) do
-    Stream.iterate(1, &(&1 + 1))
-    |> Enum.reduce_while([], fn i, acc ->
-      new_pos = {row + (i * row_dir), col + (i * col_dir)}
-      
-      cond do
-        !Map.has_key?(cells, new_pos) ->
-          {:halt, acc}
-        cells[new_pos] == nil ->
-          {:cont, [new_pos | acc]}
-        cells[new_pos].color != color ->
-          {:halt, [new_pos | acc]}
-        true ->
-          {:halt, acc}
-      end
+  defp get_diagonal_jumping_moves(cells, {row, col}, color) do
+    directions = [{1, 1}, {1, -1}, {-1, 1}, {-1, -1}]
+    
+    Enum.flat_map(directions, fn {dr, dc} = dir ->
+      Logger.info("Phoenix checking direction #{inspect(dir)}")
+      moves = get_phoenix_line(cells, {row, col}, {dr, dc}, color)
+      Logger.info("Phoenix moves in direction #{inspect(dir)}: #{inspect(moves)}")
+      moves
     end)
+  end
+
+  defp get_line_moves(cells, {row, col}, {dr, dc}, color, acc \\ []) do
+    new_pos = {row + dr, col + dc}
+    
+    cond do
+      !valid_position?(new_pos) -> 
+        acc
+      cells[new_pos] == nil ->
+        get_line_moves(cells, new_pos, {dr, dc}, color, [new_pos | acc])
+      cells[new_pos].color != color -> 
+        [new_pos | acc]
+      true -> 
+        acc
+    end
+  end
+
+  defp get_phoenix_line(cells, {row, col}, {dr, dc}, color, acc \\ []) do
+    new_pos = {row + dr, col + dc}
+    
+    cond do
+      !valid_position?(new_pos) ->
+        Enum.reverse(acc)
+      true ->
+        case cells[new_pos] do
+          nil ->
+            get_phoenix_line(cells, new_pos, {dr, dc}, color, [new_pos | acc])
+          piece when piece.color != color ->
+            get_phoenix_line(cells, new_pos, {dr, dc}, color, [new_pos | acc])
+          _piece ->
+            get_phoenix_line(cells, new_pos, {dr, dc}, color, acc)
+        end
+    end
+  end
+
+  defp valid_position?({row, col}) do
+    row >= 0 && row < 8 && col >= 0 && col < 8
   end
 end
